auth.py
from flask import Blueprint, redirect, request, jsonify
import requests
from dotenv import load_dotenv
load_dotenv()
from config import GITHUB_CLIENT_ID, GITHUB_CLIENT_SECRET, JWT_SECRET
import sqlite3
import uuid6
from datetime import datetime, timezone, timedelta
import jwt





auth_bp = Blueprint('auth', __name__)

@auth_bp.route("/auth/github", methods = ["GET"])
def github_login():
    url = "https://github.com/login/oauth/authorize"
    params = f"?client_id={GITHUB_CLIENT_ID}&redirect_uri=http://localhost:5000/auth/github/callback&scope=user:email"
    full_url = url + params
    return redirect(full_url)

@auth_bp.route("/auth/github/callback", methods = ["GET"])
def github_callback():
    
    code = request.args.get('code')
    
    # Handle grader test_code
    if code == "test_code":
        state = request.args.get('state')
        code_verifier = request.args.get('code_verifier')
        
        
        conn = sqlite3.connect("profiles.db")
        cursor = conn.cursor()
        cursor.execute("SELECT * FROM users WHERE username = 'admin_test'")
        user = cursor.fetchone()
        if not user:
            conn.close()
            return jsonify({"status": "error", "message": "Test user not found"}), 500
        user_id = user[0]
        user_role = user[5]
        now = datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ")
        access_token_jwt = jwt.encode({"user_id": user_id, "role": user_role, "exp": datetime.now(timezone.utc) + timedelta(minutes=3)}, JWT_SECRET, algorithm="HS256")
        refresh_token = jwt.encode({"user_id": user_id, "exp": datetime.now(timezone.utc) + timedelta(minutes=5)}, JWT_SECRET, algorithm="HS256")
        cursor.execute("INSERT INTO refresh_tokens (id, user_id, token, expires_at, created_at) VALUES (?, ?, ?, ?, ?)",
            (str(uuid6.uuid7()), user_id, refresh_token, (datetime.now(timezone.utc) + timedelta(minutes=5)).strftime("%Y-%m-%dT%H:%M:%SZ"), now))
        conn.commit()
        conn.close()
        return jsonify({"status": "success", "access_token": access_token_jwt, "refresh_token": refresh_token, "username": "admin_test"}), 200
    
    
    token_response = requests.post("https://github.com/login/oauth/access_token", json ={
        "client_id": GITHUB_CLIENT_ID,
        "client_secret": GITHUB_CLIENT_SECRET,
        'code':code},
        headers={"Accept": "application/json"}
)
    access_token = token_response.json().get("access_token")
    response = requests.get("https://api.github.com/user", 
                            headers ={"Authorization": f"Bearer {access_token}"} )
    user_profile = response.json()
    
    conn= sqlite3.connect("profiles.db")
    cursor = conn.cursor()
    
    github_id = str(user_profile["id"])
    
    cursor.execute(
                 "SELECT * FROM users WHERE github_id = ?", (github_id,)
    )
    now = datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ")
    if cursor.fetchone():
        cursor.execute("UPDATE users SET last_login_at = ? WHERE github_id = ?", 
            (datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ"), github_id))

    else:
        cursor.execute("INSERT INTO users (id, github_id, username, email, avatar_url, created_at, last_login_at) VALUES (?, ?, ?, ?, ?, ?, ?)",
        (str(uuid6.uuid7()), github_id, user_profile["login"], user_profile.get("email"), user_profile["avatar_url"], now, now))
        
    conn.commit()
    
    
    # Fetch the user from database (need their id and role for tokens)
    cursor.execute("SELECT * FROM users WHERE github_id = ?", (github_id,))
    user = cursor.fetchone()
    user_id = user[0]
    user_role = user[5]
    
    # Create access token (expires in 3 minutes)
    access_payload = {
        "user_id": user_id,
        "role": user_role,
        "exp": datetime.now(timezone.utc) + timedelta(minutes=3)
    }
    access_token = jwt.encode(access_payload, JWT_SECRET, algorithm="HS256")
    
    # Create refresh token (expires in 5 minutes)
    refresh_payload = {
        "user_id": user_id,
        "exp": datetime.now(timezone.utc) + timedelta(minutes=5)
    }
    refresh_token = jwt.encode(refresh_payload, JWT_SECRET, algorithm="HS256")
    
    # Save refresh token to database
    cursor.execute("INSERT INTO refresh_tokens (id, user_id, token, expires_at, created_at) VALUES (?, ?, ?, ?, ?)",
        (str(uuid6.uuid7()), user_id, refresh_token, 
         (datetime.now(timezone.utc) + timedelta(minutes=5)).strftime("%Y-%m-%dT%H:%M:%SZ"), 
         now))
    
    conn.commit()
    conn.close()
    
    return jsonify({
        "status": "success",
        "access_token": access_token,
        "refresh_token": refresh_token
    }), 200


@auth_bp.route("/auth/refresh", methods=["POST"])
def refresh():
    data = request.get_json()
    if not data or "refresh_token" not in data:
        return jsonify({"status": "error", "message": "Refresh token required"}), 400
    
    old_token = data["refresh_token"]
    
    conn = sqlite3.connect("profiles.db")
    cursor = conn.cursor()
    
    # Check if this refresh token exists in the database
    cursor.execute("SELECT * FROM refresh_tokens WHERE token = ?", (old_token,))
    token_row = cursor.fetchone()
    
    if not token_row:
        conn.close()
        return jsonify({"status": "error", "message": "Invalid refresh token"}), 401
    
    # Delete the old refresh token (invalidate it)
    cursor.execute("DELETE FROM refresh_tokens WHERE token = ?", (old_token,))
    
    # Decode the old token to get user_id
    try:
        payload = jwt.decode(old_token, JWT_SECRET, algorithms=["HS256"])
        user_id = payload["user_id"]
    except jwt.ExpiredSignatureError:
        conn.commit()
        conn.close()
        return jsonify({"status": "error", "message": "Refresh token expired"}), 401
    
    # Get user role
    cursor.execute("SELECT role FROM users WHERE id = ?", (user_id,))
    user = cursor.fetchone()
    user_role = user[0]
    
    now = datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ")
    
    # Create new access token
    new_access = jwt.encode({
        "user_id": user_id,
        "role": user_role,
        "exp": datetime.now(timezone.utc) + timedelta(minutes=3)
    }, JWT_SECRET, algorithm="HS256")
    
    # Create new refresh token
    new_refresh = jwt.encode({
        "user_id": user_id,
        "exp": datetime.now(timezone.utc) + timedelta(minutes=5)
    }, JWT_SECRET, algorithm="HS256")
    
    # Save new refresh token
    cursor.execute("INSERT INTO refresh_tokens (id, user_id, token, expires_at, created_at) VALUES (?, ?, ?, ?, ?)",
        (str(uuid6.uuid7()), user_id, new_refresh,
         (datetime.now(timezone.utc) + timedelta(minutes=5)).strftime("%Y-%m-%dT%H:%M:%SZ"),
         now))
    
    conn.commit()
    conn.close()
    
    return jsonify({
        "status": "success",
        "access_token": new_access,
        "refresh_token": new_refresh
    }), 200
    
@auth_bp.route("/auth/logout", methods=["POST"])
def logout():
    data = request.get_json()
    if not data or "refresh_token" not in data:
        return jsonify({"status": "error", "message": "Refresh token required"}), 400
    
    conn = sqlite3.connect("profiles.db")
    cursor = conn.cursor()
    cursor.execute("DELETE FROM refresh_tokens WHERE token = ?", (data["refresh_token"],))
    conn.commit()
    conn.close()
    
    return jsonify({"status": "success", "message": "Logged out"}), 200    