requirements.txt
flask
flask-cors
requests
uuid6
gunicorn
python-dotenv
PyJWT
Flask-Limiter